CSN: CSN-asimcore-321
Apply patches from Intel version, revisions 2966-3333.

CSN: CSN-simcore-3107
          

- fixed a basic error in clockserver; nsteps, which counts off base cycles,
  must match the base cycle type (UINT64).

CSN: CSN-simcore-3230
          
Eliminate an icc warning.

CSN: CSN-simcore-3291
          

- fixed a permission error in public install of simcore package.

CSN: CSN-simcore-3307
          

- a bunch of "char *" to "const char *" conversions, provided by Shadi Khasawneh,
  to get rid of warnings/errors in GCC 4.5 build

+++ lib/libasim/include/asim/clockserver.h
@@ -561,7 +561,7 @@ class ClockRegistry
      * Number that says the number of base cycles that must
      * pass before all the modules are clocked.
      **/
-    UINT32 nStep;
+    UINT64 nStep;
 
     /**
      * Says which is the NEXT base cycle that the modules are going
